+++ src/index.ts
@@ -6,6 +6,10 @@ import * as path from 'path';
 
 (function main() {
 
+    interface ITemplateVariables {
+        [index: string]: string;
+    }
+
     const packageName = process.argv[2];
 
     if (!packageName) {
@@ -25,68 +29,25 @@ import * as path from 'path';
 
     fs.mkdirSync(packageName);
 
+    const variables: ITemplateVariables = {packageName};
+
     const templateDirectory = path.join(__dirname, '..', 'templates');
 
-    // Create editorconfig
-    const editorConfigTemplate = fs.readFileSync(path.join(templateDirectory, '.editorconfig.mustache')).toString();
-    // No rendering needed
-    fs.writeFileSync(path.join(packageName, '.editorconfig'), editorConfigTemplate);
+    // The main render function.
+    function renderTemplate(templateName: string): void {
+        const template = fs.readFileSync(path.join(templateDirectory, templateName)).toString();
+        const rendered = Mustache.render(template, variables);
+        const fileName = templateName.replace('.mustache', '');
+        fs.writeFileSync(path.join(variables.packageName, fileName), rendered);
+    }
 
-    // Create code-style-settings file
-    const codeStyleSettingsTemplate = fs.readFileSync(path.join(templateDirectory, 'code-style-settings.xml.mustache')).toString();
-    const codeStyleSettingsRendered = Mustache.render(codeStyleSettingsTemplate, {packageName});
-    fs.writeFileSync(path.join(packageName, 'code-style-settings.xml'), codeStyleSettingsRendered);
+    // Render all the templates in the template folder.
+    const templates = fs.readdirSync(templateDirectory);
+    for (const template of templates) {
+        renderTemplate(template);
+    }
 
-    // Create .npmrc
-    const npmrcTemplate = fs.readFileSync(path.join(templateDirectory, '.npmrc.mustache')).toString();
-    // No rendering needed
-    fs.writeFileSync(path.join(packageName, '.npmrc'), npmrcTemplate);
-
-    // Create tsconfig
-    const tsConfigTemplate = fs.readFileSync(path.join(templateDirectory, 'tsconfig.json.mustache')).toString();
-    // No rendering needed
-    fs.writeFileSync(path.join(packageName, 'tsconfig.json'), tsConfigTemplate);
-
-    // Create tslint
-    const tsLintTemplate = fs.readFileSync(path.join(templateDirectory, 'tslint.json.mustache')).toString();
-    // No rendering needed
-    fs.writeFileSync(path.join(packageName, 'tslint.json'), tsLintTemplate);
-
-    // create gitignore
-    const gitIgnoreTemplate = fs.readFileSync(path.join(templateDirectory, '.gitignore.mustache')).toString();
-    // No rendering needed
-    fs.writeFileSync(path.join(packageName, '.gitignore'), gitIgnoreTemplate);
-
-    // Create package.json
-    const packageJsonTemplate = fs.readFileSync(path.join(templateDirectory, 'package.json.mustache')).toString();
-    const packageJsonRendered = Mustache.render(packageJsonTemplate, {packageName});
-    fs.writeFileSync(path.join(packageName, 'package.json'), packageJsonRendered);
-
-    // Create travis.yml
-    const travisTemplate = fs.readFileSync(path.join(templateDirectory, '.travis.yml.mustache')).toString();
-    // No rendering needed
-    fs.writeFileSync(path.join(packageName, '.travis.yml'), travisTemplate);
-
-    // Create LICENSE
-    const licenseTemplate = fs.readFileSync(path.join(templateDirectory, 'LICENSE.mustache')).toString();
-    // No rendering needed
-    fs.writeFileSync(path.join(packageName, 'LICENSE'), licenseTemplate);
-
-    // Create README.md
-    const readmeTemplate = fs.readFileSync(path.join(templateDirectory, 'README.md.mustache')).toString();
-    const readmeRendered = Mustache.render(readmeTemplate, {packageName});
-    fs.writeFileSync(path.join(packageName, 'README.md'), readmeRendered);
-
-    // Create CHANGELOG.md
-    const changelogTemplate = fs.readFileSync(path.join(templateDirectory, 'CHANGELOG.md.mustache')).toString();
-    const changelogRendered = Mustache.render(changelogTemplate, {packageName});
-    fs.writeFileSync(path.join(packageName, 'CHANGELOG.md'), changelogRendered);
-
-    // Create CONTRIBUTING.md
-    const contributingTemplate = fs.readFileSync(path.join(templateDirectory, 'CONTRIBUTING.md.mustache')).toString();
-    const contributingRendered = Mustache.render(contributingTemplate, {packageName});
-    fs.writeFileSync(path.join(packageName, 'CONTRIBUTING.md'), contributingRendered);
-
+    // Create standard folders.
     fs.mkdirSync(path.join(packageName, 'src'));
     fs.mkdirSync(path.join(packageName, 'dist'));