+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Badge link to package in README
 
 ### Changed
+- Rewrote render code into reusable function
 - Remove comments in dist files
 - Build for es6